chores: move tx & statistic page into dashboard, apply authGuard & tenantAuthGuard
