--- src/app/tenant/statistics/page.tsx
@@ -1,12 +1,0 @@
-import LandingPageLayout from "@/components/LandingPageLayout";
-import StatisticPage from "@/features/tenant/statistics";
-
-const Statistics = () => {
-  return (
-    <LandingPageLayout>
-      <StatisticPage />
-    </LandingPageLayout>
-  );
-};
-
-export default Statistics;
