--- src/app/tenant/transactions/page.tsx
@@ -1,12 +1,0 @@
-import LandingPageLayout from "@/components/LandingPageLayout";
-import TransactionListTenantPage from "@/features/tenant/transactions";
-
-const TenantTransactionsList = () => {
-  return (
-    <LandingPageLayout>
-      <TransactionListTenantPage />
-    </LandingPageLayout>
-  );
-};
-
-export default TenantTransactionsList;
