+++ src/app/transactions/[id]/page.tsx
@@ -1,5 +1,6 @@
 import LandingPageLayout from "@/components/LandingPageLayout";
 import TransactionDetailPage from "@/features/transactions/TransactionDetailPage";
+import UserAuthGuard from "@/hoc/userAuthGuard";
 import React from "react";
 
 const TransactionDetail = ({ params }: { params: { id: string } }) => {
@@ -10,4 +11,4 @@ const TransactionDetail = ({ params }: { params: { id: string } }) => {
   );
 };
 
-export default TransactionDetail;
+export default UserAuthGuard(TransactionDetail);
