+++ src/features/dashboard/statistics/components/MetricsChart.tsx
