+++ src/features/dashboard/statistics/components/TopPropertiesTable.tsx
